Added mouse mode
Added a boolean to encounter control to control how the player hand is traversed. If the boolean is true, then the mouse hover dictates which card is selected. Otherwise, the keyboard arrows move within the player hand.

+++ Assets/Prefabs/TimeSlot.cs
@@ -38,7 +38,7 @@ public class TimeSlot : MonoBehaviour
 
     public AbstractCard occupyingCard;
     //Start this slot's timer based on the provided cards cost
-    public IEnumerator wait(int sec, AbstractPlayer user, AbstractCard selectedCard)
+    public IEnumerator wait(float sec, AbstractPlayer user, AbstractCard selectedCard)
     {
         //Make the slot occupied
         occupied = true;
